+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Changed in upcoming release
 
+* Mobs in the broker tooltip now get click actions and informational tooltips
+    * Click a line to open the map with that mob focused
+    * Control-click a line to set a waypoint for the closest location for that mob
+    * Shift-click a line to link the closest location for that mob in chat
 * Minimap pins update their focus state immediately
 
 ## Changed in v90001.5